# DAPA.Api/wwwroot/swagger.yaml
openapi: 3.0.0
info:
  title: POS system
  description: API for managing Discounts, Services, Products, Carts, Orders, Payments, Clients, Loyalties, Reservations, Staff, and Roles
  version: 1.0.0
paths:
...
  /orders:
    get:
      tags:
        - Orders
      summary: Get orders based on filter criteria
      parameters:
        - name: id
          in: query
          description: Filter orders by ID
          schema:
            type: integer
        - name: clientId
          in: query
          description: Filter orders by client ID
          schema:
            type: integer
        - name: creationDate
          in: query
          description: Filter orders by creation date
          schema:
            type: string
            format: date-time
        - name: status
          in: query
          description: Filter orders by status
          schema:
            $ref: "#/components/schemas/Status"
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '400':
          description: Bad request
  /order:
    post:
      tags:
        - Orders
      summary: Create a new order
      requestBody:
        description: Order object to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Bad request
  /order/{id}:
    get:
      tags:
        - Orders
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the order
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Bad request

    put:
      tags:
        - Orders
      summary: Update an order by ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the order
          schema:
            type: integer
      requestBody:
        description: Updated order object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Bad request

    delete:
      tags:
        - Orders
      summary: Delete an order by ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the order
          schema:
            type: integer
      responses:
        '204':
          description: Successful response
        '400':
          description: Bad request
  /order/total/{id}:
    get:
      tags:
        - Orders
      summary: Get total based on order
      parameters:
        - name: id
          in: path
          required: true
          description: Filter orders by ID
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '400':
          description: Bad request
...
  /workinghours:
    get:
      tags:
        - Workinghours
      summary: Get all working hours
      parameters:
        - name: id
          in: query
          description: Filter working hours by ID
          schema:
            type: integer
        - name: staffid
          in: query
          description: Filter working hours by staff
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        '400':
          description: Bad request
  /workinghour:
    post:
      tags:
        - Workinghours
      summary: Create a new working hour
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkingHours'
      responses:
        '201':
          description: Working hour created successfully
        '400':
          description: Invalid input
  /workinghours/{id}:
    get:
      tags:
        - Workinghours
      summary: Get a specific working hour by ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the working hour
          schema:
            type: integer
      responses:
        '200':
          description: Successful operation
        '404':
          description: Working hour not found
    put:
      tags:
        - Workinghours
      summary: Update a working hour by ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the working hour
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkingHours'
      responses:
        '200':
          description: Working hour updated successfully
        '400':
          description: Invalid input
        '404':
          description: Working hour not found

    delete:
      tags:
        - Workinghours
      summary: Delete a working hour by ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the working hour
          schema:
            type: integer
      responses:
        '204':
          description: Working hour deleted successfully
        '404':
          description: Working hour not found
  /workinghours/available/{staffId}:
      get:
        tags:
          - Workinghours
        summary: Get available working hours for staffId
        parameters:
          - name: staffId
            in: path
            required: true
            description: ID of the staff member
            schema:
              type: integer
        responses:
          '200':
            description: Successful operation
          '404':
            description: Available staf working hours not found
components:
  schemas:
...
    Order:
      type: object
      properties:
        Id:
          type: integer
        ClientId:
          type: integer
        CreationDate:
          type: string
          format: date-time
        Status:
          $ref: "#/components/schemas/Status"
        StaffId:
          type: integer
        DiscountId:
          type: integer
        Details:
          type: string
        Tip:
          type: number
    Status:
      type: string
      enum:
        - Cancelled
        - Completed
        - InProgress
        - Paid
...
    Role:
      type: object
      properties:
        Id:
          type: integer
        Title:
          type: string
        PermissionViewOrder:
          type: boolean
        PermissionManageOrder:
          type: boolean
        PermissionCreateDiscount:
          type: boolean
        PermissionManageItems:
          type: boolean
        PermissionManageServices:
          type: boolean
        IsOwner:
          type: boolean
    WorkingHours:
      type: object
      properties:
        Id:
          type: integer
        StaffId:
          type: integer
        StartTime:
          type: string
          format: date-time
        EndTime:
          type: string
          format: date-time